Revert "Update build.yml"
This reverts commit 28def0647d6a26b8afa1c28a67a2c1b8a6813beb.

Revert "使用chafa.py代替通过subprocess调用chafa"

This reverts commit 8ca9d5c32e260ec5c2890d99eac3219d9e2e194a.

Revert "Update README.md"

This reverts commit 4f27c3f7762d3bc9df8038d5d4df3bc2a13fb2d6.

Revert "Update build.yml"

This reverts commit 829729b749a07faf46f52b79a7941fc171dacac2.

Revert "修复打包版x2"

This reverts commit 218c3600665e9fe76b4eba2b4f7f7dae9a6bc8c0.

Revert "修复Windows打包版"

This reverts commit 586c8bb5e7924e754a9600f76b8e57fc6316a56f.

Revert "Update build.yml"

This reverts commit 12ee41b205ceaaacecae4f027aadd40c3f2822fb.

Revert "Update build.yml"

This reverts commit a9869829fc285a8c81d24bd4541e62fcdf9a7ae3.

+++ xdnmb/util.py
@@ -1,11 +1,11 @@
-import chafa
-import chafa.loader
+from __future__ import annotations
 import datetime
 import functools
 import gzip
 import os
 import re
 import secrets
+import subprocess
 import tempfile
 import time
 import typing
@@ -156,6 +156,31 @@ def focusToButton(focusFrom: xdnmb.model.ButtonType | None,
         pass
     return False
 
+
+# Chafa: Terminal Graphics for the 21st Century
+# https://hpjansson.org/chafa/
+
+
+@functools.cache
+def detectChafa() -> tuple[int, int, int] | None:
+    try:
+        p = subprocess.Popen(
+            ('chafa', '--version'),
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+        p.wait()
+        if p.returncode:
+            return None
+        else:
+            s = re.search(r'^Chafa version (\d+).(\d+).(\d+)$', p.stdout.readline().decode('utf-8'))
+            if s is None:
+                s = re.search(r'^Chafa version (\d+).(\d+).(\d+)$', p.stderr.readline().decode('utf-8'))
+            return (int(s.group(1)), int(s.group(2)), int(s.group(3)))
+    except FileNotFoundError:
+        return False
+
+
 @functools.lru_cache(256)
 def loadChafaImage(url: str, width: int, height: int) -> str:
     cacheKey = ':'.join((url, str(width), str(height)))
@@ -163,33 +188,44 @@ def loadChafaImage(url: str, width: int, height: int) -> str:
     if cached:
         return gzip.decompress(cached).decode('utf-8')
 
+    # Reading image data from stdin raises "Failed to open '-': Unknown file format" error on Windows build · Issue #100 · hpjansson/chafa
+    # https://github.com/hpjansson/chafa/issues/100
+    # Fixed in Chafa 1.12.4
+    # https://github.com/hpjansson/chafa/releases/tag/1.12.4
+    useTemp: bool = os.name == 'nt' and detectChafa() < (1, 12, 4)
     temp = os.path.join(tempfile.gettempdir(), secrets.token_urlsafe(12))
-    with (
-        xdnmb.api.session.get(url, stream=True, timeout=3) as r,
-        open(temp, 'wb') as f
-    ):
-        for chunk in r.iter_content(8192):
-            f.write(chunk)
-
-    image = chafa.loader.Loader(temp)
-
-    config = chafa.CanvasConfig()
-    config.width = width
-    config.height = height
-    config.work_factor = 1
-    config.optimizations = (chafa.Optimizations.CHAFA_OPTIMIZATION_ALL,)
-    config.calc_canvas_geometry(image.width, image.height, .5)
-
-    canvas = chafa.Canvas(config)
-    canvas.draw_all_pixels(
-        image.pixel_type,
-        image.get_pixels(),
-        image.width, image.height,
-        image.rowstride,
+    cmd = (
+        'chafa',
+        '--duration',
+        str(0),
+        '--optimize',
+        str(9),
+        '--size',
+        f'{width}x{height}',
+        '--work',
+        str(9),
+        temp if useTemp else '-',
     )
-
-    result = canvas.print()
-
-    os.remove(temp)
-    lruCacheSet(cacheKey, gzip.compress(result, 9), 16384)
-    return result.decode()
+    with xdnmb.api.session.get(url, stream=True, timeout=3) as r:
+        if useTemp:
+            with open(temp, 'wb') as f:
+                for chunk in r.iter_content(8192):
+                    f.write(chunk)
+        with subprocess.Popen(
+            cmd,
+            stdin=subprocess.DEVNULL if useTemp else subprocess.PIPE,
+            stdout=subprocess.PIPE,
+            creationflags=subprocess.CREATE_NO_WINDOW if os.name == 'nt' else 0,
+        ) as p:
+            if not useTemp:
+                for chunk in r.iter_content(8192):
+                    p.stdin.write(chunk)
+                p.stdin.close()
+            result = p.stdout.read()
+        if useTemp:
+            os.remove(temp)
+    if p.returncode:
+        raise subprocess.CalledProcessError(p.returncode, cmd)
+    result = re.split(r'\033\[\d*A', result.decode('utf-8'), 1)[0].replace('\r', '').strip()
+    lruCacheSet(cacheKey, gzip.compress(result.encode('utf-8'), 9), 16384)
+    return result
